feat(store): add analytics section and refocus orders on operations
Introduce /dashboard/store/analytics for charts, stats, low-stock alerts, and
product totals. Move operational summary cards onto orders with links to
payments and products. Extract StoreNav, show a pending-voucher indicator on
payments, and load pending counts in the server layout. Default /dashboard/store
and the Tienda nav entry to analytics; revalidate analytics when orders change.
